ActivityLoaderApp/requirements.txt
azure-functions
pymongo
dnspython